Remove Myocardium_demo module
